+++ src/main/java/frc/robot/Robot.java
@@ -6,6 +6,9 @@ package frc.robot;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
+import frc.robot.subsystems.elevator.Elevator;
+import frc.robot.subsystems.elevator.ElevatorConstants;
+import frc.robot.subsystems.elevator.factories.ElevatorFactory;
 
 /**
  * This class is where the bulk of the robot should be declared. Since Command-based is a "declarative" paradigm, very little robot logic should
@@ -14,9 +17,14 @@ import edu.wpi.first.wpilibj2.command.InstantCommand;
  */
 public class Robot {
 
+	private final Elevator elevator;
+
 	public static final RobotType ROBOT_TYPE = RobotType.determineRobotType();
 
 	public Robot() {
+		this.elevator = new Elevator(ElevatorFactory.create(ElevatorConstants.LOG_PATH));
+		// add bindings here
+
 		configureBindings();
 	}
